# Switch to gunicorn's `gthread` worker type