gunicorn's default worker type is the `sync` worker, which is more suited to CPU/network-bandwidth bound workloads.

As such this switches to the thread-based `gthread` worker for improved request throughput performance for blocking I/O workloads - given that it's common for Heroku apps to make blocking requests to external APIs/datastores etc.

The threads count of 5 is somewhat arbitrary, but seemed to work well enough with rough benchmarking against a few different dyno sizes (where the process count will vary according to `WEB_CONCURRENCY`) and simulated workload types.

The `preload_app` option has also been enabled, which makes gunicorn load the app before the worker processes are forked, to reduce memory usage and boot times.

See:
https://docs.gunicorn.org/en/stable/design.html#server-model
https://docs.gunicorn.org/en/stable/settings.html#worker-class

GUS-W-17236632.