@edmorley 2024-11-21
Using:

* A new endpoint with an artificial 200ms sleep (as an example of a slow external blocking call):

    ```python
    # views.py
    import time
    
    def sleep(request):
        time.sleep(0.2)
        return render(request, "index.html")
    ```
* 2 x Standard-1X dynos on Common Runtime
* The [default `WEB_CONCURRENCY`](https://devcenter.heroku.com/articles/python-concurrency) set by the Python buildpack
* `http-routing-2-dot-0` enabled
* VPN switched off
* https://github.com/wg/wrk

## Before

#### With 20 clients:

```
$ wrk -t 6 -c 20 -d 10s https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
Running 10s test @ https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
  6 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   882.00ms  400.19ms   1.60s    61.38%
    Req/Sec     3.72      2.62    10.00     66.86%
  189 requests in 10.10s, 1.50MB read
Requests/sec:     18.71
Transfer/sec:    152.20KB
```

#### With 40 clients:

```
$ wrk -t 6 -c 40 -d 10s https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
Running 10s test @ https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
  6 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.50s   392.39ms   1.91s    81.20%
    Req/Sec     3.71      2.70    14.00     67.55%
  192 requests in 10.09s, 1.53MB read
  Socket errors: connect 0, read 0, write 0, timeout 59
Requests/sec:     19.02
Transfer/sec:    154.78KB
```

## After

#### With 20 clients:

```
$ wrk -t 6 -c 20 -d 10s https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
Running 10s test @ https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
  6 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   292.61ms   21.88ms 491.03ms   92.62%
    Req/Sec    10.08      3.98    20.00     74.80%
  596 requests in 10.09s, 4.73MB read
Requests/sec:     59.07
Transfer/sec:    480.31KB
```

#### With 40 clients:

```
$ wrk -t 6 -c 40 -d 10s https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
Running 10s test @ https://emorley-concurrency-test-22abc062cbe4.herokuapp.com/sleep/
  6 threads and 40 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   384.06ms   90.34ms 684.25ms   48.12%
    Req/Sec    15.79      8.22    50.00     80.37%
  902 requests in 10.09s, 7.16MB read
Requests/sec:     89.43
Transfer/sec:    727.23KB
```